Renamed modals reducer module.
