+++ src/js/rx/reducers/index.js
@@ -2,7 +2,7 @@ import { combineReducers } from 'redux';
 
 import alert from './alert';
 import disclaimer from './disclaimer';
-import modals from './modal';
+import modals from './modals';
 import prescriptions from './prescriptions';
 
 export default combineReducers({
