+++ src/js/rx/reducers/modals.js
